rename `"babel-plugin"` package to `"transform"`
