+++ .changeset/happy-falcons-smash.md
@@ -1,0 +1,5 @@
+---
+"@solid-devtools/transform": minor
+---
+
+Add wrapStores transform.
