+++ packages/locator/package.json
@@ -1,78 +1,78 @@
 {
-	"name": "@solid-devtools/locator",
-	"version": "0.8.1",
-	"description": "On-page overlay for localisating SolidJS components, and revealing them in your code editor.",
-	"license": "MIT",
-	"author": "Damian Tarnawski <gthetarnav@gmail.com>",
-	"contributors": [],
-	"homepage": "https://github.com/thetarnav/solid-devtools/tree/main/packages/locator#readme",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/thetarnav/solid-devtools.git"
-	},
-	"bugs": {
-		"url": "https://github.com/thetarnav/solid-devtools/issues"
-	},
-	"keywords": [
-		"solid",
-		"devtools",
-		"components",
-		"locator",
-		"vscode"
-	],
-	"private": false,
-	"sideEffects": false,
-	"type": "module",
-	"files": [
-		"dist"
-	],
-	"main": "./dist/server.cjs",
-	"module": "./dist/server.js",
-	"types": "./dist/index.d.ts",
-	"exports": {
-		"browser": {
-			"development": {
-				"import": "./dist/index.js",
-				"require": "./dist/index.cjs"
-			},
-			"import": "./dist/server.js",
-			"require": "./dist/server.cjs"
-		},
-		"import": "./dist/server.js",
-		"require": "./dist/server.cjs"
-	},
-	"scripts": {
-		"dev": "tsup --watch",
-		"build": "tsup",
-		"test": "jest --config jest.config.ts",
-		"typecheck": "tsc --noEmit"
-	},
-	"devDependencies": {
-		"@testing-library/jest-dom": "^5.16.4",
-		"esbuild": "^0.14.51",
-		"esbuild-plugin-solid": "^0.4.2",
-		"jest": "^28.1.3",
-		"jest-environment-jsdom": "^28.1.3",
-		"solid-js": "^1.4.8",
-		"ts-node": "^10.9.1",
-		"tsup": "^6.2.1",
-		"typescript": "^4.7.4"
-	},
-	"dependencies": {
-		"@solid-devtools/debugger": "^0.2.3",
-		"@solid-devtools/ui": "0.3.3",
-		"@solid-primitives/bounds": "^0.0.100",
-		"@solid-primitives/event-listener": "^2.2.1",
-		"@solid-primitives/keyboard": "^0.0.100",
-		"@solid-primitives/platform": "^0.0.100",
-		"@solid-primitives/utils": "^3.0.1",
-		"motion": "^10.13.1"
-	},
-	"peerDependencies": {
-		"solid-js": "^1.4.4"
-	},
-	"optionalDependencies": {
-		"@solid-devtools/babel-plugin": "^0.3.1"
-	},
-	"packageManager": "pnpm@7.7.0"
+  "name": "@solid-devtools/locator",
+  "version": "0.8.1",
+  "description": "On-page overlay for localisating SolidJS components, and revealing them in your code editor.",
+  "license": "MIT",
+  "author": "Damian Tarnawski <gthetarnav@gmail.com>",
+  "contributors": [],
+  "homepage": "https://github.com/thetarnav/solid-devtools/tree/main/packages/locator#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/thetarnav/solid-devtools.git"
+  },
+  "bugs": {
+    "url": "https://github.com/thetarnav/solid-devtools/issues"
+  },
+  "keywords": [
+    "solid",
+    "devtools",
+    "components",
+    "locator",
+    "vscode"
+  ],
+  "private": false,
+  "sideEffects": false,
+  "type": "module",
+  "files": [
+    "dist"
+  ],
+  "main": "./dist/server.cjs",
+  "module": "./dist/server.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    "browser": {
+      "development": {
+        "import": "./dist/index.js",
+        "require": "./dist/index.cjs"
+      },
+      "import": "./dist/server.js",
+      "require": "./dist/server.cjs"
+    },
+    "import": "./dist/server.js",
+    "require": "./dist/server.cjs"
+  },
+  "scripts": {
+    "dev": "tsup --watch",
+    "build": "tsup",
+    "test": "jest --config jest.config.ts",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.4",
+    "esbuild": "^0.14.51",
+    "esbuild-plugin-solid": "^0.4.2",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "solid-js": "^1.4.8",
+    "ts-node": "^10.9.1",
+    "tsup": "^6.2.1",
+    "typescript": "^4.7.4"
+  },
+  "dependencies": {
+    "@solid-devtools/debugger": "^0.2.3",
+    "@solid-devtools/ui": "0.3.3",
+    "@solid-primitives/bounds": "^0.0.100",
+    "@solid-primitives/event-listener": "^2.2.1",
+    "@solid-primitives/keyboard": "^0.0.100",
+    "@solid-primitives/platform": "^0.0.100",
+    "@solid-primitives/utils": "^3.0.1",
+    "motion": "^10.13.1"
+  },
+  "peerDependencies": {
+    "solid-js": "^1.4.4"
+  },
+  "optionalDependencies": {
+    "@solid-devtools/transform": "^0.3.1"
+  },
+  "packageManager": "pnpm@7.7.0"
 }
