+++ packages/main/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - fdb09bc: Various minor changes.
 - Updated dependencies [fdb09bc]
-  - @solid-devtools/babel-plugin@0.3.1
+  - @solid-devtools/transform@0.3.1
   - @solid-devtools/debugger@0.2.3
   - @solid-devtools/extension-adapter@0.8.1
   - @solid-devtools/locator@0.8.1
@@ -20,7 +20,7 @@
 ### Patch Changes
 
 - Updated dependencies [4c79a90]
-  - @solid-devtools/babel-plugin@0.3.0
+  - @solid-devtools/transform@0.3.0
   - @solid-devtools/locator@0.8.0
 
 ## 0.7.2
@@ -29,7 +29,7 @@
 
 - a8d0354: Correct "homepage" filed in package.json, to lead to individual package readme.
 - Updated dependencies [a8d0354]
-  - @solid-devtools/babel-plugin@0.2.1
+  - @solid-devtools/transform@0.2.1
   - @solid-devtools/debugger@0.2.2
   - @solid-devtools/extension-adapter@0.7.2
   - @solid-devtools/locator@0.7.2
