+++ packages/main/README.md
@@ -11,7 +11,7 @@ The main package of Solid Devtools. It contains the following subpackages:
 - [Debugger](https://github.com/thetarnav/solid-devtools/tree/main/packages/debugger#readme)
 - [Extension Adapter](https://github.com/thetarnav/solid-devtools/tree/main/packages/extension-adapter#readme) _(this one is automatically enabled)_
 - [Locator](https://github.com/thetarnav/solid-devtools/tree/main/packages/locator#readme) — [How to use it](#using-the-locator-package)
-- [Babel Plugin](https://github.com/thetarnav/solid-devtools/tree/main/packages/babel-plugin#readme) — [How to use it](#enabling-the-babel-plugin)
+- [Babel Plugin](https://github.com/thetarnav/solid-devtools/tree/main/packages/transform#readme) — [How to use it](#enabling-the-babel-plugin)
 
 ## Getting started
 
@@ -33,7 +33,7 @@ The `solid-devtools` package comes with the [Locator](https://github.com/thetarn
 
 ### Enabling the Babel plugin
 
-`solid-devtools` reexports the [babel plugin](https://github.com/thetarnav/solid-devtools/tree/main/packages/babel-plugin#readme) as a vite plugin.
+`solid-devtools` reexports the [babel plugin](https://github.com/thetarnav/solid-devtools/tree/main/packages/transform#readme) as a vite plugin.
 
 To enable it you need to add it to plugins array in your `.vite.config.js` file:
 
@@ -48,6 +48,8 @@ export default defineConfig({
   plugins: [devtoolsPlugin(), solidPlugin()],
 })
 ```
+
+[**See transform options**](https://github.com/thetarnav/solid-devtools/tree/main/packages/transform#Options)
 
 ## Changelog
