+++ packages/main/package.json
@@ -1,85 +1,85 @@
 {
-	"name": "solid-devtools",
-	"version": "0.8.1",
-	"description": "Runtime helpers for hooking up SolidJS application with Solid Devtools",
-	"license": "MIT",
-	"author": "Damian Tarnawski <gthetarnav@gmail.com>",
-	"contributors": [],
-	"homepage": "https://github.com/thetarnav/solid-devtools/tree/main/packages/main#readme",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/thetarnav/solid-devtools.git"
-	},
-	"bugs": {
-		"url": "https://github.com/thetarnav/solid-devtools/issues"
-	},
-	"keywords": [
-		"solid",
-		"devtools",
-		"reactivity"
-	],
-	"private": false,
-	"sideEffects": true,
-	"publishConfig": {
-		"access": "public"
-	},
-	"files": [
-		"dist"
-	],
-	"type": "module",
-	"main": "./dist/server.cjs",
-	"module": "./dist/server.js",
-	"types": "./dist/index.d.ts",
-	"exports": {
-		".": {
-			"browser": {
-				"development": {
-					"import": "./dist/index.js",
-					"require": "./dist/index.cjs"
-				},
-				"import": "./dist/server.js",
-				"require": "./dist/server.cjs"
-			},
-			"import": "./dist/server.js",
-			"require": "./dist/server.cjs",
-			"types": "./dist/index.d.ts"
-		},
-		"./vite": {
-			"import": "./dist/vite.js",
-			"require": "./dist/vite.cjs",
-			"types": "./dist/vite.d.ts"
-		}
-	},
-	"typesVersions": {
-		"*": {
-			"vite": [
-				"./dist/vite.d.ts"
-			]
-		}
-	},
-	"scripts": {
-		"dev": "tsup --watch",
-		"build": "tsup",
-		"test": "jest --config jest.config.ts",
-		"typecheck": "tsc --noEmit"
-	},
-	"devDependencies": {
-		"@testing-library/jest-dom": "^5.16.4",
-		"jest": "^28.1.3",
-		"jest-environment-jsdom": "^28.1.3",
-		"solid-js": "^1.4.8",
-		"ts-node": "^10.9.1",
-		"tsup": "^6.2.1",
-		"typescript": "^4.7.4"
-	},
-	"dependencies": {
-		"@solid-devtools/babel-plugin": "^0.3.1",
-		"@solid-devtools/debugger": "^0.2.3",
-		"@solid-devtools/extension-adapter": "^0.8.1",
-		"@solid-devtools/locator": "^0.8.1"
-	},
-	"peerDependencies": {
-		"solid-js": "^1.4.5"
-	},
-	"packageManager": "pnpm@7.7.0"
+  "name": "solid-devtools",
+  "version": "0.8.1",
+  "description": "Runtime helpers for hooking up SolidJS application with Solid Devtools",
+  "license": "MIT",
+  "author": "Damian Tarnawski <gthetarnav@gmail.com>",
+  "contributors": [],
+  "homepage": "https://github.com/thetarnav/solid-devtools/tree/main/packages/main#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/thetarnav/solid-devtools.git"
+  },
+  "bugs": {
+    "url": "https://github.com/thetarnav/solid-devtools/issues"
+  },
+  "keywords": [
+    "solid",
+    "devtools",
+    "reactivity"
+  ],
+  "private": false,
+  "sideEffects": true,
+  "publishConfig": {
+    "access": "public"
+  },
+  "files": [
+    "dist"
+  ],
+  "type": "module",
+  "main": "./dist/server.cjs",
+  "module": "./dist/server.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "browser": {
+        "development": {
+          "import": "./dist/index.js",
+          "require": "./dist/index.cjs"
+        },
+        "import": "./dist/server.js",
+        "require": "./dist/server.cjs"
+      },
+      "import": "./dist/server.js",
+      "require": "./dist/server.cjs",
+      "types": "./dist/index.d.ts"
+    },
+    "./vite": {
+      "import": "./dist/vite.js",
+      "require": "./dist/vite.cjs",
+      "types": "./dist/vite.d.ts"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      "vite": [
+        "./dist/vite.d.ts"
+      ]
+    }
+  },
+  "scripts": {
+    "dev": "tsup --watch",
+    "build": "tsup",
+    "test": "jest --config jest.config.ts",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.4",
+    "jest": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "solid-js": "^1.4.8",
+    "ts-node": "^10.9.1",
+    "tsup": "^6.2.1",
+    "typescript": "^4.7.4"
+  },
+  "dependencies": {
+    "@solid-devtools/transform": "^0.3.1",
+    "@solid-devtools/debugger": "^0.2.3",
+    "@solid-devtools/extension-adapter": "^0.8.1",
+    "@solid-devtools/locator": "^0.8.1"
+  },
+  "peerDependencies": {
+    "solid-js": "^1.4.5"
+  },
+  "packageManager": "pnpm@7.7.0"
 }
