+++ packages/ui/package.json
@@ -1,51 +1,51 @@
 {
-	"name": "@solid-devtools/ui",
-	"version": "0.3.3",
-	"description": "UI Library for visualization of the reactive graph. Powering Solid Devtools.",
-	"license": "MIT",
-	"author": "Damian Tarnawski <gthetarnav@gmail.com>",
-	"contributors": [],
-	"homepage": "https://github.com/thetarnav/solid-devtools/tree/main/packages/ui#readme",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/thetarnav/solid-devtools.git"
-	},
-	"bugs": {
-		"url": "https://github.com/thetarnav/solid-devtools/issues"
-	},
-	"keywords": [
-		"solid",
-		"devtools"
-	],
-	"private": false,
-	"sideEffects": false,
-	"type": "module",
-	"main": "dist/index.cjs",
-	"module": "dist/index.js",
-	"types": "dist/index.d.ts",
-	"scripts": {
-		"dev": "tsup --watch",
-		"build": "tsup",
-		"test": "echo \"NOOP TEST\"",
-		"typecheck": "tsc --noEmit"
-	},
-	"devDependencies": {
-		"@solid-devtools/babel-plugin": "^0.3.1",
-		"esbuild": "^0.14.51",
-		"esbuild-plugin-solid": "^0.4.2",
-		"solid-js": "^1.4.8",
-		"tsup": "^6.2.1",
-		"typescript": "^4.7.4"
-	},
-	"dependencies": {
-		"@otonashixav/solid-flip": "^0.10.5",
-		"@solid-aria/interactions": "^0.1.4",
-		"@solid-primitives/props": "^2.2.1",
-		"@solid-primitives/utils": "^3.0.1",
-		"twind": "^0.16.17"
-	},
-	"peerDependencies": {
-		"solid-js": "^1.4.4"
-	},
-	"packageManager": "pnpm@7.7.0"
+  "name": "@solid-devtools/ui",
+  "version": "0.3.3",
+  "description": "UI Library for visualization of the reactive graph. Powering Solid Devtools.",
+  "license": "MIT",
+  "author": "Damian Tarnawski <gthetarnav@gmail.com>",
+  "contributors": [],
+  "homepage": "https://github.com/thetarnav/solid-devtools/tree/main/packages/ui#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/thetarnav/solid-devtools.git"
+  },
+  "bugs": {
+    "url": "https://github.com/thetarnav/solid-devtools/issues"
+  },
+  "keywords": [
+    "solid",
+    "devtools"
+  ],
+  "private": false,
+  "sideEffects": false,
+  "type": "module",
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "dev": "tsup --watch",
+    "build": "tsup",
+    "test": "echo \"NOOP TEST\"",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@solid-devtools/transform": "^0.3.1",
+    "esbuild": "^0.14.51",
+    "esbuild-plugin-solid": "^0.4.2",
+    "solid-js": "^1.4.8",
+    "tsup": "^6.2.1",
+    "typescript": "^4.7.4"
+  },
+  "dependencies": {
+    "@otonashixav/solid-flip": "^0.10.5",
+    "@solid-aria/interactions": "^0.1.4",
+    "@solid-primitives/props": "^2.2.1",
+    "@solid-primitives/utils": "^3.0.1",
+    "twind": "^0.16.17"
+  },
+  "peerDependencies": {
+    "solid-js": "^1.4.4"
+  },
+  "packageManager": "pnpm@7.7.0"
 }
