+++ playgrounds/sandbox/package.json
@@ -1,26 +1,26 @@
 {
-	"name": "solid-devtools-sandbox-playground",
-	"version": "0.0.1",
-	"private": true,
-	"description": "",
-	"scripts": {
-		"start": "vite",
-		"dev": "vite",
-		"build": "vite build",
-		"serve": "vite preview"
-	},
-	"license": "MIT",
-	"devDependencies": {
-		"@solid-devtools/babel-plugin": "^0.3.1",
-		"typescript": "^4.7.4",
-		"vite": "^3.0.4",
-		"vite-plugin-solid": "^2.3.0"
-	},
-	"dependencies": {
-		"@solid-devtools/logger": "^0.2.0",
-		"@solid-primitives/timer": "^1.3.1",
-		"object-observer": "^5.0.4",
-		"solid-devtools": "^0.8.1",
-		"solid-js": "^1.4.8"
-	}
+  "name": "solid-devtools-sandbox-playground",
+  "version": "0.0.1",
+  "private": true,
+  "description": "",
+  "scripts": {
+    "start": "vite",
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview"
+  },
+  "license": "MIT",
+  "devDependencies": {
+    "@solid-devtools/transform": "^0.3.1",
+    "typescript": "^4.7.4",
+    "vite": "^3.0.4",
+    "vite-plugin-solid": "^2.3.0"
+  },
+  "dependencies": {
+    "@solid-devtools/logger": "^0.2.0",
+    "@solid-primitives/timer": "^1.3.1",
+    "object-observer": "^5.0.4",
+    "solid-devtools": "^0.8.1",
+    "solid-js": "^1.4.8"
+  }
 }
