+++ playgrounds/start/package.json
@@ -1,28 +1,28 @@
 {
-	"name": "solid-devtools-start-playground",
-	"version": "0.0.1",
-	"private": true,
-	"scripts": {
-		"dev": "solid-start dev",
-		"build": "solid-start build",
-		"start": "solid-start start"
-	},
-	"type": "module",
-	"main": "./dist/index.js",
-	"devDependencies": {
-		"@solid-devtools/babel-plugin": "^0.3.1",
-		"@solid-devtools/logger": "^0.2.0",
-		"solid-app-router": "^0.4.2",
-		"solid-devtools": "^0.8.1",
-		"solid-js": "^1.4.8",
-		"solid-meta": "^0.27.5",
-		"solid-start": "next",
-		"solid-start-node": "next",
-		"typescript": "^4.7.4",
-		"undici": "^4.16.0",
-		"vite": "2.9.12"
-	},
-	"engines": {
-		"node": ">=14"
-	}
+  "name": "solid-devtools-start-playground",
+  "version": "0.0.1",
+  "private": true,
+  "scripts": {
+    "dev": "solid-start dev",
+    "build": "solid-start build",
+    "start": "solid-start start"
+  },
+  "type": "module",
+  "main": "./dist/index.js",
+  "devDependencies": {
+    "@solid-devtools/transform": "^0.3.1",
+    "@solid-devtools/logger": "^0.2.0",
+    "solid-app-router": "^0.4.2",
+    "solid-devtools": "^0.8.1",
+    "solid-js": "^1.4.8",
+    "solid-meta": "^0.27.5",
+    "solid-start": "next",
+    "solid-start-node": "next",
+    "typescript": "^4.7.4",
+    "undici": "^4.16.0",
+    "vite": "2.9.12"
+  },
+  "engines": {
+    "node": ">=14"
+  }
 }
