+++ playgrounds/start/vite.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "vite"
 import solid from "solid-start"
-import { devtoolsPlugin } from "@solid-devtools/babel-plugin"
+import { devtoolsPlugin } from "@solid-devtools/transform"
 
 export default defineConfig({
-	plugins: [devtoolsPlugin(), solid()],
+  plugins: [devtoolsPlugin(), solid()],
 })
